+++ packages/base/src/features/story/presentation/index.ts
@@ -1,0 +1,10 @@
+export type { StoryPresentationMode } from './types';
+export type { IStoryPresentationRootProps, IStoryStageProps } from './types';
+
+export {
+  getStoryPresentationMode,
+  isVerticalScrollPresentation,
+} from './getStoryPresentationMode';
+
+export { StoryStage } from './StoryStage';
+export { StoryPresentationRoot } from './StoryPresentationRoot';
