+++ packages/base/src/features/story/presentation/modes/column/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Column presentation chrome (guided).
+ * SpectaDesktopView / SpectaMobileView single branches migrate here.
+ */
+export {};
