+++ packages/base/src/features/story/presentation/modes/verticalScroll/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Vertical-scroll presentation chrome.
+ * ListStoryVirtualScrollTrack / SpectaMobileListModeContent migrate here.
+ */
+export {};
